pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "pyaudio>=0.2.14",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "wave>=0.0.2",
    "websocket-client>=1.8.0",
]
